requirements.txt
# shared
psycopg2-binary>=2.9.9

# analyzer
claude-agent-sdk>=0.1.0
feedparser>=6.0.11
httpx>=0.27.0
anyio>=4.4.0
yfinance>=0.2.0

# api
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0

# auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1
